Create new .swift files and change models destinations
