+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Роман on 29.12.2022.
+//
+
+import Foundation
+
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
