+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Роман on 29.12.2022.
+//
+
+import Foundation
+
+struct QuizResultViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
