+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Роман on 29.12.2022.
+//
+
+import UIKit
+
+struct QuizStepViewModel {
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
